Upgrade upstream to v1.20.2 (#134)
* refactoring to disable most AppExtrinsic related tests

* update deps

* fix duplicate deps

* update polkadot-sdk deps

* fix features

* deep cleaning

+++ core/src/data_proof.rs
@@ -141,8 +141,8 @@ impl DataProof {
 			root,
 			proof: self.proof.clone(),
 			leaf,
-			number_of_leaves: self.number_of_leaves as usize,
-			leaf_index: self.leaf_index as usize,
+			number_of_leaves: self.number_of_leaves,
+			leaf_index: self.leaf_index,
 		}
 	}
 }
